refactor: 拆分 teams.ts 为模块化结构
**问题：**
teams.ts 1315 行，超过 Cursor 1000 行代码质量红线

**改造方案：**
- 按职责分层拆分为 6 个模块
- 每个文件 < 600 行

**新结构：**
api/src/routes/teams/
├── index.ts        # 路由聚合入口 (42 行)
├── queries.ts      # 查询类路由 (522 行) - GET
├── mutations.ts    # 写入类路由 (255 行) - POST/PUT/DELETE
├── membership.ts   # 成员管理 (323 行) - 加入/审批/移除
├── status.ts       # 状态变更 (174 行) - 成行/取消/退出
└── utils.ts        # 工具函数 (89 行)

**原文件兼容：**
teams.ts 改为从 teams/index.ts 重新导出，保持向后兼容

**测试：**
- TypeScript 编译通过
- 路由路径保持不变

Refs: 遵循 Cursor thermo-nuclear-code-quality-review 原则
