+++ api/src/routes/teams.ts
@@ -1,1315 +1,14 @@
-import { Hono } from "hono";
-import { eq, desc, and, ne, sql, like, inArray, lt, gte, lte } from "drizzle-orm";
-import { createAuth } from "../lib/auth";
-import { createDb } from "../db";
-import * as schema from "../db/schema";
-import type { Env } from "../lib/auth";
-import { sendTeamJoinApplicationEmail } from "../lib/email";
-
-const teams = new Hono<{ Bindings: Env }>();
-
-/** 随机队伍图标 */
-const TEAM_ICONS = ["⛰️", "🥾", "🌲", "🏕️", "🧗", "🌄", "🏞️", "🗺️"];
-function getRandomTeamIcon() {
-  return TEAM_ICONS[Math.floor(Math.random() * TEAM_ICONS.length)];
-}
-
-/**
- * 获取时间快捷筛选的日期范围（北京时间）
- */
-function getTimeFilterRange(timeFilter: string): { start: string; end: string } | null {
-  const BEIJING_OFFSET_MS = 8 * 60 * 60 * 1000;
-  const now = new Date();
-  const beijingNow = new Date(now.getTime() + BEIJING_OFFSET_MS);
-
-  const formatDate = (d: Date) => d.toISOString().split("T")[0];
-
-  switch (timeFilter) {
-    case "today": {
-      const dateStr = formatDate(beijingNow);
-      return { start: dateStr, end: dateStr };
-    }
-    case "tomorrow": {
-      const tomorrow = new Date(beijingNow);
-      tomorrow.setDate(tomorrow.getDate() + 1);
-      const dateStr = formatDate(tomorrow);
-      return { start: dateStr, end: dateStr };
-    }
-    case "weekend": {
-      const dayOfWeek = beijingNow.getUTCDay(); // 0=Sunday, 6=Saturday
-      const daysUntilSaturday = dayOfWeek === 0 ? 6 : 6 - dayOfWeek;
-
-      const saturday = new Date(beijingNow);
-      saturday.setDate(beijingNow.getDate() + daysUntilSaturday);
-
-      const sunday = new Date(saturday);
-      sunday.setDate(saturday.getDate() + 1);
-
-      return { start: formatDate(saturday), end: formatDate(sunday) };
-    }
-    case "7days": {
-      const start = formatDate(beijingNow);
-      const endDate = new Date(beijingNow);
-      endDate.setDate(endDate.getDate() + 7);
-      return { start, end: formatDate(endDate) };
-    }
-    default:
-      return null;
-  }
-}
-
-/**
- * 安全解析 requirements 字段
- * 旧数据可能是非 JSON 字符串，解析失败返回空数组
- */
-function parseRequirements(value: string | null | undefined): string[] {
-  if (!value) return [];
-  try {
-    const parsed = JSON.parse(value);
-    return Array.isArray(parsed) ? parsed : [];
-  } catch {
-    return [];
-  }
-}
-
-/** 获取单个路线的标签 */
-async function getRouteTags(
-  db: ReturnType<typeof createDb>,
-  routeId: string
-): Promise<{ id: string; name: string; type: string }[]> {
-  const tagResults = await db
-    .select({ tag: schema.tags })
-    .from(schema.entityToTags)
-    .leftJoin(schema.tags, eq(schema.entityToTags.tagId, schema.tags.id))
-    .where(
-      and(
-        eq(schema.entityToTags.entityId, routeId),
-        eq(schema.entityToTags.entityType, "route")
-      )
-    );
-
-  return tagResults
-    .filter((result) => result.tag)
-    .map((result) => ({
-      id: result.tag!.id,
-      name: result.tag!.name,
-      type: result.tag!.type,
-    }));
-}
-
-/** 将已过期的队伍更新为 completed */
-async function updateExpiredTeams(db: ReturnType<typeof createDb>, teamId?: string) {
-  const now = new Date();
-
-  // 1. 将 endTime 已过期的 formed 队伍更新为 completed
-  const formedCondition = teamId
-    ? and(eq(schema.teams.id, teamId), eq(schema.teams.status, "formed"), lt(schema.teams.endTime, now))
-    : and(eq(schema.teams.status, "formed"), lt(schema.teams.endTime, now));
-  await db
-    .update(schema.teams)
-    .set({ status: "completed", updatedAt: now })
-    // eslint-disable-next-line @typescript-eslint/no-explicit-any
-    .where(formedCondition as any);
-
-  // 2. 将 startTime 已过期的 recruiting 队伍更新为 completed
-  const recruitingCondition = teamId
-    ? and(eq(schema.teams.id, teamId), eq(schema.teams.status, "recruiting"), lt(schema.teams.startTime, now))
-    : and(eq(schema.teams.status, "recruiting"), lt(schema.teams.startTime, now));
-  await db
-    .update(schema.teams)
-    .set({ status: "completed", updatedAt: now })
-    // eslint-disable-next-line @typescript-eslint/no-explicit-any
-    .where(recruitingCondition as any);
-}
-
-/**
- * GET /teams
- * 获取队伍列表，支持分页、搜索、status/difficulty 过滤
- * 也支持 locationId / userId+includeJoined / activeOnly 特殊模式
- */
-teams.get("/", async (c) => {
-  try {
-    const db = createDb(c.env.DB);
-
-    // 特殊模式参数
-    const locationId = c.req.query("locationId");
-    const userId = c.req.query("userId");
-    const includeJoined = c.req.query("includeJoined") === "true";
-    const activeOnly = c.req.query("activeOnly") === "true";
-
-    // 列表模式参数
-    const page = Math.max(1, parseInt(c.req.query("page") || "1", 10));
-    const pageSize = Math.min(100, parseInt(c.req.query("pageSize") || "12", 10));
-    const search = c.req.query("search") || "";
-    const statusParam = c.req.query("status") || "";
-    const difficultyParam = c.req.query("difficulty") || "";
-
-    // 日期范围筛选参数
-    const startDateFrom = c.req.query("startDateFrom"); // ISO 日期格式 YYYY-MM-DD
-    const startDateTo = c.req.query("startDateTo");     // ISO 日期格式 YYYY-MM-DD
-
-    // 时间快捷筛选参数（today/tomorrow/weekend/7days）
-    const timeFilter = c.req.query("timeFilter");
-
-    // 标签筛选参数
-    const tagIdsParam = c.req.query("tagIds") || "";
-
-    await updateExpiredTeams(db);
-
-    const currentMembersSubquery = sql<number>`(
-      SELECT COUNT(*) FROM team_members
-      WHERE team_members.team_id = ${schema.teams.id}
-      AND team_members.status IN ('approved', 'leave_pending')
-    )`;
-
-    const teamColumns = {
-      id: schema.teams.id,
-      locationId: schema.teams.locationId,
-      routeId: schema.teams.routeId,
-      leaderId: schema.teams.leaderId,
-      title: schema.teams.title,
-      description: schema.teams.description,
-      startTime: schema.teams.startTime,
-      endTime: schema.teams.endTime,
-      durationMin: schema.teams.durationMin,
-      maxMembers: schema.teams.maxMembers,
-      requirements: schema.teams.requirements,
-      icon: schema.teams.icon,
-      status: schema.teams.status,
-      createdAt: schema.teams.createdAt,
-      updatedAt: schema.teams.updatedAt,
-      currentMembers: currentMembersSubquery,
-      leaderImage: schema.users.image,
-      leaderName: schema.users.name,
-      leaderNickname: schema.users.nickname,
-      leaderLevel: schema.users.level,
-      locationName: schema.locations.name,
-      locationCoverImage: schema.locations.coverImage,
-    };
-
-    type TeamRow = {
-      id: string; locationId: string; routeId: string | null; leaderId: string;
-      title: string; description: string | null; startTime: Date; endTime: Date;
-      durationMin: number | null; maxMembers: number; requirements: string | null;
-      icon: string; status: string; createdAt: Date; updatedAt: Date;
-      currentMembers: number; leaderImage: string | null; leaderName: string;
-      leaderNickname: string | null; leaderLevel: string | null;
-      locationName: string | null; locationCoverImage: string | null;
-    };
-
-    let result: TeamRow[];
-
-    // 特殊模式：我加入的队伍
-    if (userId && includeJoined) {
-      const conditions = [
-        eq(schema.teamMembers.userId, userId),
-        ne(schema.teams.leaderId, userId),
-        eq(schema.teamMembers.status, "approved"),
-      ];
-      if (activeOnly) {
-        conditions.push(ne(schema.teams.status, "completed"));
-        conditions.push(ne(schema.teams.status, "cancelled"));
-      }
-      result = await db
-        .select(teamColumns)
-        .from(schema.teams)
-        .innerJoin(schema.teamMembers, eq(schema.teamMembers.teamId, schema.teams.id))
-        .innerJoin(schema.users, eq(schema.users.id, schema.teams.leaderId))
-        .leftJoin(schema.locations, eq(schema.locations.id, schema.teams.locationId))
-        .where(and(...conditions))
-        .orderBy(desc(schema.teams.createdAt)) as TeamRow[];
-      return c.json({ success: true, teams: formatTeams(result) });
-    }
-
-    // 特殊模式：某地点的队伍
-    if (locationId) {
-      const conditions = [eq(schema.teams.locationId, locationId)];
-
-      // 添加 status 过滤
-      if (statusParam) {
-        const statuses = statusParam.split(",").filter(Boolean);
-        if (statuses.length === 1) {
-          conditions.push(eq(schema.teams.status, statuses[0]));
-        } else if (statuses.length > 1) {
-          conditions.push(inArray(schema.teams.status, statuses));
-        }
-      }
-
-      result = await db
-        .select(teamColumns)
-        .from(schema.teams)
-        .innerJoin(schema.users, eq(schema.users.id, schema.teams.leaderId))
-        .leftJoin(schema.locations, eq(schema.locations.id, schema.teams.locationId))
-        .where(and(...conditions))
-        .orderBy(desc(schema.teams.createdAt)) as TeamRow[];
-      return c.json({ success: true, teams: formatTeams(result) });
-    }
-
-    // 通用列表模式：支持搜索、status、difficulty、日期范围、标签、分页
-    const conditions = [];
-
-    if (statusParam) {
-      const statuses = statusParam.split(",").filter(Boolean);
-      if (statuses.length === 1) {
-        conditions.push(eq(schema.teams.status, statuses[0]));
-      } else if (statuses.length > 1) {
-        conditions.push(inArray(schema.teams.status, statuses));
-      }
-    }
-
-    if (search) {
-      conditions.push(like(schema.teams.title, `%${search}%`));
-    }
-
-    // 日期范围筛选（支持 timeFilter 快捷参数或 startDateFrom/To 自定义范围）
-    let effectiveStartDate = startDateFrom;
-    let effectiveEndDate = startDateTo;
-
-    // 如果提供了 timeFilter，计算对应的日期范围
-    if (timeFilter && !effectiveStartDate && !effectiveEndDate) {
-      const range = getTimeFilterRange(timeFilter);
-      if (range) {
-        effectiveStartDate = range.start;
-        effectiveEndDate = range.end;
-      }
-    }
-
-    if (effectiveStartDate) {
-      const fromDate = new Date(effectiveStartDate);
-      fromDate.setHours(0, 0, 0, 0);
-      conditions.push(gte(schema.teams.startTime, fromDate));
-    }
-    if (effectiveEndDate) {
-      const toDate = new Date(effectiveEndDate);
-      toDate.setHours(23, 59, 59, 999);
-      conditions.push(lte(schema.teams.startTime, toDate));
-    }
-
-    // difficulty 过滤需要 join routes 表
-    const difficultyList = difficultyParam ? difficultyParam.split(",").filter(Boolean) : [];
-    
-    // 标签筛选
-    const tagIds = tagIdsParam ? tagIdsParam.split(",").filter(Boolean) : [];
-
-    const whereClause = conditions.length > 0 ? and(...conditions) : undefined;
-
-    // 如果有标签筛选，先查出符合条件的 teamIds
-    let filteredTeamIds: string[] | null = null;
-    if (tagIds.length > 0) {
-      const tagResults = await db
-        .select({ entityId: schema.entityToTags.entityId })
-        .from(schema.entityToTags)
-        .where(
-          and(
-            eq(schema.entityToTags.entityType, "activity"),
-            inArray(schema.entityToTags.tagId, tagIds)
-          )
-        );
-      filteredTeamIds = tagResults.map(r => r.entityId);
-      if (filteredTeamIds.length === 0) {
-        // 没有符合条件的队伍，直接返回空结果
-        return c.json({
-          success: true,
-          teams: [],
-          pagination: { page, pageSize, total: 0, totalPages: 0 },
-        });
-      }
-    }
-
-    // 构建最终 where 条件
-    const finalWhereClause = (() => {
-      const allConditions = [];
-      if (whereClause) allConditions.push(whereClause);
-      if (difficultyList.length > 0) {
-        allConditions.push(inArray(schema.routes.difficulty, difficultyList));
-      }
-      if (filteredTeamIds) {
-        allConditions.push(inArray(schema.teams.id, filteredTeamIds));
-      }
-      return allConditions.length > 0 ? and(...allConditions) : undefined;
-    })();
-
-    // 统计总数
-    const [{ cnt }] = await db
-      .select({ cnt: sql<number>`count(distinct ${schema.teams.id})` })
-      .from(schema.teams)
-      .leftJoin(schema.routes, eq(schema.routes.id, schema.teams.routeId))
-      .where(finalWhereClause);
-    const total = cnt;
-
-    const totalPages = Math.ceil(total / pageSize);
-    const offset = (page - 1) * pageSize;
-
-    // 查询列表
-    result = await db
-      .select(teamColumns)
-      .from(schema.teams)
-      .innerJoin(schema.users, eq(schema.users.id, schema.teams.leaderId))
-      .leftJoin(schema.locations, eq(schema.locations.id, schema.teams.locationId))
-      .leftJoin(schema.routes, eq(schema.routes.id, schema.teams.routeId))
-      .where(finalWhereClause)
-      .orderBy(desc(schema.teams.startTime))
-      .limit(pageSize)
-      .offset(offset) as TeamRow[];
-
-    return c.json({
-      success: true,
-      teams: formatTeams(result),
-      pagination: { page, pageSize, total, totalPages },
-    });
-  } catch (error) {
-    console.error("Get teams error:", error);
-    return c.json({ error: "获取队伍列表失败" }, 500);
-  }
-});
-
-/** 格式化队伍行数据为前端所需格式 */
-function formatTeams(result: {
-  id: string; locationId: string; routeId: string | null; leaderId: string;
-  title: string; description: string | null; startTime: Date; endTime: Date;
-  durationMin: number | null; maxMembers: number; requirements: string | null;
-  icon: string; status: string; createdAt: Date; updatedAt: Date;
-  currentMembers: number; leaderImage: string | null; leaderName: string;
-  leaderNickname: string | null; leaderLevel: string | null;
-  locationName: string | null; locationCoverImage: string | null;
-}[]) {
-  return result.map((row) => {
-    const startDate = new Date(row.startTime);
-    const endDate = new Date(row.endTime);
-    // 使用北京时间（UTC+8）格式化日期和时间
-    const beijingDate = new Date(startDate.getTime() + 8 * 60 * 60 * 1000);
-    const date = beijingDate.toISOString().split("T")[0];
-    const time = beijingDate.toISOString().slice(11, 16);
-    const durationHours = Math.round(
-      (endDate.getTime() - startDate.getTime()) / (1000 * 60 * 60)
-    );
-    const requirements = parseRequirements(row.requirements);
-    return {
-      id: row.id, locationId: row.locationId, routeId: row.routeId,
-      title: row.title, description: row.description || "",
-      date, time, duration: `${durationHours}小时`,
-      durationMin: row.durationMin || durationHours * 60,
-      maxMembers: row.maxMembers, currentMembers: row.currentMembers,
-      icon: row.icon || "⭿️", requirements, status: row.status,
-      createdAt: row.createdAt,
-      location: row.locationName ? {
-        name: row.locationName,
-        coverImage: row.locationCoverImage || "",
-      } : undefined,
-      leader: {
-        id: row.leaderId, name: row.leaderName,
-        nickname: row.leaderNickname || null,
-        avatar: row.leaderImage || "",
-        level: (row.leaderLevel || "beginner") as string,
-        completedHikes: 0, bio: "",
-      },
-    };
-  });
-}
-
-/**
- * POST /teams
- * 创建新队伍（需登录，需填写微信号）
- */
-teams.post("/", async (c) => {
-  try {
-    const authInstance = createAuth(c.env);
-    const session = await authInstance.api.getSession({ headers: c.req.raw.headers });
-    if (!session) return c.json({ error: "请先登录" }, 401);
-
-    const db = createDb(c.env.DB);
-    const userId = session.user.id;
-
-    const userRecord = await db
-      .select({ wechat: schema.users.wechat })
-      .from(schema.users)
-      .where(eq(schema.users.id, userId))
-      .then((rows) => rows[0]);
-
-    if (!userRecord?.wechat) {
-      return c.json({ error: "请先填写微信号才能创建队伍" }, 400);
-    }
-
-    const body = await c.req.json<{
-      locationId?: string; routeId?: string; title?: string;
-      description?: string; date?: string; time?: string;
-      duration?: string; durationMin?: number; maxMembers?: number;
-      requirements?: string[];
-    }>();
-    const { locationId, routeId, title, description, date, time, duration, durationMin, maxMembers, requirements } = body;
-
-    if (!locationId || !title || !date || !time || !maxMembers) {
-      return c.json({ error: "缺少必填字段" }, 400);
-    }
-
-    const startTime = new Date(`${date}T${time}`);
-    if (isNaN(startTime.getTime())) {
-      return c.json({ error: "无效的日期或时间格式" }, 400);
-    }
-
-    const durationMinutes = durationMin || (duration
-      ? parseFloat(duration.replace(/[^0-9.]/g, "")) * 60 || 240
-      : 240);
-    const endTime = new Date(startTime.getTime() + durationMinutes * 60 * 1000);
-    const teamId = `team-${Date.now()}-${Math.random().toString(36).substr(2, 9)}`;
-    const memberId = `tm-${Date.now()}-${Math.random().toString(36).substr(2, 9)}`;
-    const now = new Date();
-    const teamIcon = getRandomTeamIcon();
-
-    await db.insert(schema.teams).values({
-      id: teamId, locationId, routeId, leaderId: userId, title,
-      description: description || null, startTime, endTime,
-      durationMin: durationMinutes, maxMembers,
-      requirements: requirements ? JSON.stringify(requirements) : null,
-      icon: teamIcon, status: "recruiting", createdAt: now, updatedAt: now,
-    });
-    await db.insert(schema.teamMembers).values({
-      id: memberId, teamId, userId, status: "approved",
-      joinedAt: now, createdAt: now,
-    });
-
-    return c.json({ success: true, team: { id: teamId, locationId, routeId, leaderId: userId, title, description, startTime: startTime.toISOString(), endTime: endTime.toISOString(), durationMin: durationMinutes, maxMembers, currentMembers: 1, requirements, icon: teamIcon, status: "recruiting", createdAt: now.toISOString() } });
-  } catch (error) {
-    console.error("Create team error:", error);
-    return c.json({ error: "创建队伍失败" }, 500);
-  }
-});
-
-/**
- * GET /teams/:id
- * 获取单个队伍详情
- */
-teams.get("/:id", async (c) => {
-  try {
-    const teamId = c.req.param("id");
-    const db = createDb(c.env.DB);
-
-    await updateExpiredTeams(db, teamId);
-
-    const teamWithRelations = await db.query.teams.findFirst({
-      where: eq(schema.teams.id, teamId),
-      with: { leader: true, members: { with: { user: true } }, route: true, location: true },
-    });
-
-    if (!teamWithRelations) return c.json({ error: "队伍不存在" }, 404);
-
-    let routeTags: { id: string; name: string; type: string }[] = [];
-    if (teamWithRelations.routeId) {
-      routeTags = await getRouteTags(db, teamWithRelations.routeId);
-    }
-
-    // 获取当前登录用户（可选）
-    let currentUserId: string | null = null;
-    try {
-      const authInstance = createAuth(c.env);
-      const session = await authInstance.api.getSession({ headers: c.req.raw.headers });
-      currentUserId = session?.user?.id || null;
-    } catch { /* 用户未登录 */ }
-
-    // 计算已加入人数：approved + leave_pending 成员
-    // 如果队长不在 members 数组中，需要 +1
-    const leaderInMembers = teamWithRelations.members?.some(
-      (m: { userId: string }) => m.userId === teamWithRelations.leaderId
-    );
-    const currentMembers = (teamWithRelations.members?.filter(
-      (m: { status: string }) => m.status === "approved" || m.status === "leave_pending"
-    ).length || 0) + (leaderInMembers ? 0 : 1);
-
-    const isTeamMember = currentUserId
-      ? !!teamWithRelations.members?.find(
-          (m: { userId: string; status: string }) =>
-            m.userId === currentUserId && m.status === "approved"
-        )
-      : false;
-
-    const startDate = new Date(teamWithRelations.startTime);
-    // 使用北京时间（UTC+8）格式化日期和时间
-    const beijingDate = new Date(startDate.getTime() + 8 * 60 * 60 * 1000);
-    const date = beijingDate.toISOString().split("T")[0];
-    const time = beijingDate.toISOString().slice(11, 16);
-    const durationMinutes = teamWithRelations.durationMin ||
-      Math.round((new Date(teamWithRelations.endTime).getTime() - startDate.getTime()) / 60000);
-    const durationHours = Math.round(durationMinutes / 60);
-
-    type MemberWithUser = {
-      userId: string; status: string; joinedAt: Date | null;
-      user: { id: string; name: string; nickname?: string | null; image: string | null;
-               bio: string | null; level: string | null; wechat: string | null;
-               gender: string | null; birthday: Date | number | null; extra: string | null };
-    };
-
-    const relevantMembers = teamWithRelations.members
-      ?.filter((m: { status: string }) => m.status === "approved" || m.status === "leave_pending")
-      .map((m: MemberWithUser) => ({
-        id: m.user?.id || m.userId,
-        userId: m.userId,
-        name: m.user?.name || "未知用户",
-        nickname: m.user?.nickname || null,
-        avatar: m.user?.image || null,
-        bio: m.user?.bio || null,
-        level: m.user?.level || "beginner",
-        status: m.status,
-        joinedAt: m.joinedAt,
-        wechat: isTeamMember ? (m.user?.wechat || undefined) : undefined,
-        gender: m.user?.gender || null,
-        birthday: m.user?.birthday || null,
-        extra: m.user?.extra || null,
-      })) || [];
-
-    type LeaderUser = { id: string; name: string; nickname?: string | null; image: string | null;
-                        bio: string | null; level: string | null; wechat: string | null;
-                        gender: string | null; birthday: Date | number | null; extra: string | null };
-    const leader = teamWithRelations.leader as LeaderUser;
-
-    return c.json({
-      success: true,
-      team: {
-        id: teamWithRelations.id, locationId: teamWithRelations.locationId,
-        routeId: teamWithRelations.routeId, title: teamWithRelations.title,
-        description: teamWithRelations.description || "", date, time,
-        duration: `${durationHours}小时`, durationMin: durationMinutes,
-        maxMembers: teamWithRelations.maxMembers, currentMembers,
-        requirements: parseRequirements(teamWithRelations.requirements),
-        status: teamWithRelations.status, createdAt: teamWithRelations.createdAt,
-        route: teamWithRelations.route
-          ? {
-              id: teamWithRelations.route.id,
-              name: teamWithRelations.route.name,
-              difficulty: teamWithRelations.route.difficulty,
-              durationMin: teamWithRelations.route.durationMin,
-              durationMax: teamWithRelations.route.durationMax,
-              distance: teamWithRelations.route.distance,
-              elevation: teamWithRelations.route.elevation,
-              tags: routeTags,
-            }
-          : undefined,
-        location: teamWithRelations.location ? {
-          id: teamWithRelations.location.id,
-          name: teamWithRelations.location.name,
-          coverImage: teamWithRelations.location.coverImage,
-        } : undefined,
-        leader: leader ? {
-          id: leader.id, name: leader.name, nickname: leader.nickname || null,
-          avatar: leader.image || "", level: leader.level || "beginner",
-          completedHikes: 0, bio: leader.bio || "",
-          wechat: (isTeamMember || currentUserId === leader.id) ? (leader.wechat || "") : undefined,
-          gender: leader.gender || null, birthday: leader.birthday || null,
-          extra: leader.extra || null,
-        } : { id: "unknown", name: "未知用户", avatar: "", level: "beginner", completedHikes: 0, bio: "" },
-        members: relevantMembers,
-      },
-    });
-  } catch (error) {
-    console.error("Get team error:", error);
-    return c.json({ error: "获取队伍详情失败" }, 500);
-  }
-});
-
-/**
- * PUT /teams/:id
- * 更新队伍信息（仅队长可操作）
- */
-teams.put("/:id", async (c) => {
-  try {
-    const authInstance = createAuth(c.env);
-    const session = await authInstance.api.getSession({ headers: c.req.raw.headers });
-    if (!session) return c.json({ success: false, error: "请先登录" }, 401);
-
-    const teamId = c.req.param("id");
-    const db = createDb(c.env.DB);
-
-    const team = await db.query.teams.findFirst({
-      where: eq(schema.teams.id, teamId),
-      with: { location: true },
-    });
-
-    if (!team) return c.json({ success: false, error: "队伍不存在" }, 404);
-    if (team.leaderId !== session.user.id)
-      return c.json({ success: false, error: "只有队长可以修改队伍" }, 403);
-
-    const body = await c.req.json<{
-      title?: string; description?: string; maxMembers?: number;
-      requirements?: string[]; time?: string; durationMin?: number;
-    }>();
-    const { title, description, maxMembers, requirements, time, durationMin } = body;
-
-    if (!title || !maxMembers) return c.json({ success: false, error: "缺少必填字段" }, 400);
-
-    type UpdateData = {
-      title: string; description: string | null; maxMembers: number;
-      requirements: string | null; updatedAt: Date;
-      durationMin?: number; startTime?: Date; endTime?: Date;
-    };
-    const updateData: UpdateData = {
-      title, description: description || null, maxMembers,
-      requirements: requirements ? JSON.stringify(requirements) : null,
-      updatedAt: new Date(),
-    };
-
-    if (time || durationMin) {
-      const originalStartTime = new Date(team.startTime);
-      const currentDurationMin = durationMin || team.durationMin || 240;
-      let newStartTime = originalStartTime;
-      if (time) {
-        const [hours, minutes] = time.split(":").map(Number);
-        // time is in Beijing (UTC+8); get the Beijing date of original startTime
-        const BEIJING_OFFSET_MS = 8 * 60 * 60 * 1000;
-        const originalInBeijing = new Date(originalStartTime.getTime() + BEIJING_OFFSET_MS);
-        const beijingMs = Date.UTC(
-          originalInBeijing.getUTCFullYear(),
-          originalInBeijing.getUTCMonth(),
-          originalInBeijing.getUTCDate(),
-          hours, minutes, 0, 0
-        );
-        newStartTime = new Date(beijingMs - BEIJING_OFFSET_MS);
-        updateData.startTime = newStartTime;
-      }
-      updateData.endTime = new Date(newStartTime.getTime() + currentDurationMin * 60000);
-      if (durationMin) updateData.durationMin = durationMin;
-    }
-
-    await db.update(schema.teams).set(updateData).where(eq(schema.teams.id, teamId));
-
-    return c.json({ success: true, message: "队伍信息已更新" });
-  } catch (error) {
-    console.error("Update team error:", error);
-    return c.json({ success: false, error: "更新队伍失败" }, 500);
-  }
-});
-
-/**
- * POST /teams/:id/join
- * 申请加入队伍
- */
-teams.post("/:id/join", async (c) => {
-  try {
-    const authInstance = createAuth(c.env);
-    const session = await authInstance.api.getSession({ headers: c.req.raw.headers });
-    if (!session) return c.json({ error: "请先登录" }, 401);
-
-    const userId = session.user.id;
-    const db = createDb(c.env.DB);
-
-    const userRecord = await db
-      .select({ wechat: schema.users.wechat })
-      .from(schema.users)
-      .where(eq(schema.users.id, userId))
-      .then((rows) => rows[0]);
-
-    if (!userRecord?.wechat) return c.json({ error: "请先填写微信号才能加入队伍" }, 400);
-
-    const teamId = c.req.param("id");
-    if (!teamId) return c.json({ error: "缺少队伍ID" }, 400);
-
-    const team = await db.query.teams.findFirst({ where: eq(schema.teams.id, teamId) });
-    if (!team) return c.json({ error: "队伍不存在" }, 404);
-    if (team.status !== "recruiting") return c.json({ error: "该队伍当前不接受新成员" }, 400);
-
-    const [{ approvedCount }] = await db
-      .select({ approvedCount: sql<number>`count(*)` })
-      .from(schema.teamMembers)
-      .where(and(eq(schema.teamMembers.teamId, teamId), eq(schema.teamMembers.status, "approved")));
-
-    if (approvedCount >= team.maxMembers) return c.json({ error: "队伍已满" }, 400);
-
-    const existingMembers = await db.query.teamMembers.findMany({
-      where: and(eq(schema.teamMembers.teamId, teamId), eq(schema.teamMembers.userId, userId)),
-      limit: 1,
-    });
-    const existing = existingMembers[0];
-
-    if (existing) {
-      if (existing.status === "approved") return c.json({ error: "你已经是该队伍的成员" }, 400);
-      if (existing.status === "pending") return c.json({ error: "你已经提交了申请，请等待审核" }, 400);
-      if (existing.status === "rejected") {
-        await db.update(schema.teamMembers)
-          .set({ status: "pending", createdAt: new Date(), statusUpdatedAt: new Date() })
-          .where(eq(schema.teamMembers.id, existing.id));
-        // 异步发送通知邮件（在 Cloudflare Workers 环境中使用 waitUntil，否则直接执行）
-        const notifyPromise = notifyLeaderOfApplication(db, team, userId, c.env).catch((err) => {
-          console.error("[Email] Team join notification failed:", err);
-        });
-        try {
-          if (c.executionCtx?.waitUntil) {
-            c.executionCtx.waitUntil(notifyPromise);
-          }
-        } catch {
-          // 非 Cloudflare 环境，直接执行不阻塞
-        }
-        return c.json({ success: true, message: "重新申请已提交" });
-      }
-    }
-
-    const memberId = `tm-${Date.now()}-${Math.random().toString(36).substr(2, 9)}`;
-    await db.insert(schema.teamMembers).values({
-      id: memberId, teamId, userId, status: "pending", createdAt: new Date(),
-    });
-    // 异步发送通知邮件（在 Cloudflare Workers 环境中使用 waitUntil，否则直接执行）
-    const notifyPromise = notifyLeaderOfApplication(db, team, userId, c.env).catch((err) => {
-      console.error("[Email] Team join notification failed:", err);
-    });
-    try {
-      if (c.executionCtx?.waitUntil) {
-        c.executionCtx.waitUntil(notifyPromise);
-      }
-    } catch {
-      // 非 Cloudflare 环境，直接执行不阻塞
-    }
-
-    return c.json({ success: true, message: "申请已提交，等待队长审核" });
-  } catch (error) {
-    console.error("Join team error:", error);
-    return c.json({ error: "申请加入失败" }, 500);
-  }
-});
-
-/**
- * GET /teams/:id/applications
- * 获取待审核申请列表（仅队长）
- */
-teams.get("/:id/applications", async (c) => {
-  try {
-    const authInstance = createAuth(c.env);
-    const session = await authInstance.api.getSession({ headers: c.req.raw.headers });
-    if (!session) return c.json({ error: "请先登录" }, 401);
-
-    const teamId = c.req.param("id");
-    const db = createDb(c.env.DB);
-
-    const team = await db.query.teams.findFirst({ where: eq(schema.teams.id, teamId) });
-    if (!team) return c.json({ error: "队伍不存在" }, 404);
-    if (team.leaderId !== session.user.id)
-      return c.json({ error: "只有队长可以查看申请列表" }, 403);
-
-    // 支持 status 查询参数过滤；不传则返回全部成员（管理页面需要）
-    const statusFilter = c.req.query("status");
-    const whereClause = statusFilter
-      ? and(eq(schema.teamMembers.teamId, teamId), eq(schema.teamMembers.status, statusFilter))
-      : eq(schema.teamMembers.teamId, teamId);
-
-    const applications = await db.query.teamMembers.findMany({
-      where: whereClause,
-      with: { user: { columns: { id: true, name: true, nickname: true, image: true, bio: true, level: true, wechat: true } } },
-      orderBy: [desc(schema.teamMembers.createdAt)],
-    });
-
-    return c.json({
-      success: true,
-      applications: applications.map((app) => ({
-        id: app.id, userId: app.userId, status: app.status, createdAt: app.createdAt,
-        userName: app.user ? (app.user.nickname || app.user.name) : "",
-        wechat: app.user?.wechat || null,
-        user: app.user ? { id: app.user.id, name: app.user.name, nickname: app.user.nickname || null, avatar: app.user.image || null, bio: app.user.bio || null, level: app.user.level || "beginner" } : null,
-      })),
-    });
-  } catch (error) {
-    console.error("Get applications error:", error);
-    return c.json({ error: "获取申请列表失败" }, 500);
-  }
-});
-
-/**
- * POST /teams/:id/members/:userId/approve
- * 批准成员申请（仅队长）
- */
-teams.post("/:id/members/:userId/approve", async (c) => {
-  try {
-    const authInstance = createAuth(c.env);
-    const session = await authInstance.api.getSession({ headers: c.req.raw.headers });
-    if (!session) return c.json({ error: "请先登录" }, 401);
-
-    const teamId = c.req.param("id");
-    const targetUserId = c.req.param("userId");
-    const db = createDb(c.env.DB);
-
-    const team = await db.query.teams.findFirst({
-      where: eq(schema.teams.id, teamId), with: { location: true },
-    });
-    if (!team) return c.json({ error: "队伍不存在" }, 404);
-    if (team.leaderId !== session.user.id) return c.json({ error: "只有队长可以审核成员" }, 403);
-
-    const members = await db.query.teamMembers.findMany({
-      where: and(eq(schema.teamMembers.teamId, teamId), eq(schema.teamMembers.userId, targetUserId), eq(schema.teamMembers.status, "pending")),
-      limit: 1,
-    });
-    const membership = members[0];
-    if (!membership) return c.json({ error: "未找到该成员的申请" }, 404);
-
-    const [{ approvedCount }] = await db
-      .select({ approvedCount: sql<number>`count(*)` })
-      .from(schema.teamMembers)
-      .where(and(eq(schema.teamMembers.teamId, teamId), eq(schema.teamMembers.status, "approved")));
-
-    if (approvedCount >= team.maxMembers) return c.json({ error: "队伍已满，无法批准新成员" }, 400);
-
-    const now = new Date();
-    const newCount = approvedCount + 1;
-    const newStatus = newCount >= team.maxMembers ? "full" : "recruiting";
-
-    await db.update(schema.teamMembers)
-      .set({ status: "approved", joinedAt: now, statusUpdatedAt: now })
-      .where(eq(schema.teamMembers.id, membership.id));
-    await db.update(schema.teams)
-      .set({ status: newStatus, updatedAt: now })
-      .where(eq(schema.teams.id, teamId));
-
-    return c.json({ success: true, message: "已通过申请" });
-  } catch (error) {
-    console.error("Approve member error:", error);
-    return c.json({ error: "批准申请失败" }, 500);
-  }
-});
-
-/**
- * POST /teams/:id/members/:userId/reject
- * 拒绝成员申请（仅队长）
- */
-teams.post("/:id/members/:userId/reject", async (c) => {
-  try {
-    const authInstance = createAuth(c.env);
-    const session = await authInstance.api.getSession({ headers: c.req.raw.headers });
-    if (!session) return c.json({ error: "请先登录" }, 401);
-
-    const teamId = c.req.param("id");
-    const targetUserId = c.req.param("userId");
-    const db = createDb(c.env.DB);
-
-    const body = await c.req.json<{ reason?: string }>().catch(() => ({} as { reason?: string }));
-    const { reason } = body;
-
-    const team = await db.query.teams.findFirst({ where: eq(schema.teams.id, teamId) });
-    if (!team) return c.json({ error: "队伍不存在" }, 404);
-    if (team.leaderId !== session.user.id) return c.json({ error: "只有队长可以审核成员" }, 403);
-
-    const members = await db.query.teamMembers.findMany({
-      where: and(eq(schema.teamMembers.teamId, teamId), eq(schema.teamMembers.userId, targetUserId), eq(schema.teamMembers.status, "pending")),
-      limit: 1,
-    });
-    const membership = members[0];
-    if (!membership) return c.json({ error: "未找到该成员的申请" }, 404);
-
-    const now = new Date();
-    const extra = reason ? JSON.stringify({ rejectReason: reason }) : null;
-
-    await db.update(schema.teamMembers)
-      .set({ status: "rejected", statusUpdatedAt: now, extra })
-      .where(eq(schema.teamMembers.id, membership.id));
-
-    // 检查并更新队伍状态：如果已满则恢复为招募中
-    if (team.status === "full") {
-      await db.update(schema.teams)
-        .set({ status: "recruiting", updatedAt: now })
-        .where(eq(schema.teams.id, teamId));
-    }
-
-    return c.json({ success: true, message: "已拒绝申请" });
-  } catch (error) {
-    console.error("Reject member error:", error);
-    return c.json({ error: "拒绝申请失败" }, 500);
-  }
-});
-
-/**
- * POST /teams/:id/leave
- * 成员退出队伍
- */
-teams.post("/:id/leave", async (c) => {
-  try {
-    const authInstance = createAuth(c.env);
-    const session = await authInstance.api.getSession({ headers: c.req.raw.headers });
-    if (!session) return c.json({ success: false, error: "请先登录" }, 401);
-
-    const teamId = c.req.param("id");
-    const db = createDb(c.env.DB);
-
-    const team = await db.query.teams.findFirst({
-      where: eq(schema.teams.id, teamId), with: { location: true },
-    });
-    if (!team) return c.json({ success: false, error: "队伍不存在" }, 404);
-
-    const members = await db.query.teamMembers.findMany({
-      where: and(eq(schema.teamMembers.teamId, teamId), eq(schema.teamMembers.userId, session.user.id), eq(schema.teamMembers.status, "approved")),
-      limit: 1,
-    });
-    const membership = members[0];
-    if (!membership) return c.json({ success: false, error: "你不是该队伍成员" }, 400);
-    if (membership.userId === team.leaderId) return c.json({ success: false, error: "队长不能退出队伍" }, 400);
-    if (team.status === "formed") return c.json({ success: false, error: "队伍已组建，请通过退出申请流程离开" }, 400);
-
-    await db.delete(schema.teamMembers).where(eq(schema.teamMembers.id, membership.id));
-
-    const [{ remainingCount }] = await db
-      .select({ remainingCount: sql<number>`count(*)` })
-      .from(schema.teamMembers)
-      .where(and(eq(schema.teamMembers.teamId, teamId), eq(schema.teamMembers.status, "approved")));
-
-    await db.update(schema.teams)
-      .set({ status: remainingCount < team.maxMembers ? "recruiting" : team.status, updatedAt: new Date() })
-      .where(eq(schema.teams.id, teamId));
-
-    return c.json({ success: true, message: "已成功退出队伍" });
-  } catch (error) {
-    console.error("Leave team error:", error);
-    return c.json({ success: false, error: "退出队伍失败" }, 500);
-  }
-});
-
-/**
- * POST /teams/:id/cancel-application
- * 取消入队申请
- */
-teams.post("/:id/cancel-application", async (c) => {
-  try {
-    const authInstance = createAuth(c.env);
-    const session = await authInstance.api.getSession({ headers: c.req.raw.headers });
-    if (!session) return c.json({ error: "请先登录" }, 401);
-
-    const teamId = c.req.param("id");
-    const db = createDb(c.env.DB);
-
-    const members = await db.query.teamMembers.findMany({
-      where: and(eq(schema.teamMembers.teamId, teamId), eq(schema.teamMembers.userId, session.user.id), eq(schema.teamMembers.status, "pending")),
-      limit: 1,
-    });
-    const membership = members[0];
-    if (!membership) return c.json({ error: "未找到待审核的申请" }, 404);
-
-    // 改为更新状态而不是删除，保留历史记录
-    await db.update(schema.teamMembers)
-      .set({ status: "cancelled", statusUpdatedAt: new Date() })
-      .where(eq(schema.teamMembers.id, membership.id));
-
-    return c.json({ success: true, message: "申请已取消" });
-  } catch (error) {
-    console.error("Cancel application error:", error);
-    return c.json({ error: "取消申请失败" }, 500);
-  }
-});
-
-/**
- * POST /teams/:id/members/:userId/remove
- * 移除成员（仅队长）
- */
-teams.post("/:id/members/:userId/remove", async (c) => {
-  try {
-    const authInstance = createAuth(c.env);
-    const session = await authInstance.api.getSession({ headers: c.req.raw.headers });
-    if (!session) return c.json({ success: false, error: "请先登录" }, 401);
-
-    const teamId = c.req.param("id");
-    const targetUserId = c.req.param("userId");
-    const db = createDb(c.env.DB);
-
-    const team = await db.query.teams.findFirst({ where: eq(schema.teams.id, teamId) });
-    if (!team) return c.json({ success: false, error: "队伍不存在" }, 404);
-    if (team.leaderId !== session.user.id) return c.json({ success: false, error: "只有队长可以移除成员" }, 403);
-    if (targetUserId === session.user.id) return c.json({ success: false, error: "不能移除自己" }, 400);
-
-    const members = await db.query.teamMembers.findMany({
-      where: and(eq(schema.teamMembers.teamId, teamId), eq(schema.teamMembers.userId, targetUserId), eq(schema.teamMembers.status, "approved")),
-      limit: 1,
-    });
-    const membership = members[0];
-    if (!membership) return c.json({ success: false, error: "该用户不是队伍成员" }, 404);
-
-    await db.delete(schema.teamMembers).where(eq(schema.teamMembers.id, membership.id));
-
-    const [{ remainingCount }] = await db
-      .select({ remainingCount: sql<number>`count(*)` })
-      .from(schema.teamMembers)
-      .where(and(eq(schema.teamMembers.teamId, teamId), eq(schema.teamMembers.status, "approved")));
-
-    await db.update(schema.teams)
-      .set({ status: remainingCount < team.maxMembers ? "recruiting" : team.status, updatedAt: new Date() })
-      .where(eq(schema.teams.id, teamId));
-
-    return c.json({ success: true, message: "已移除成员" });
-  } catch (error) {
-    console.error("Remove member error:", error);
-    return c.json({ success: false, error: "移除成员失败" }, 500);
-  }
-});
-
-/**
- * POST /teams/:id/leave-request
- * 申请退出已组建的队伍
- */
-teams.post("/:id/leave-request", async (c) => {
-  try {
-    const authInstance = createAuth(c.env);
-    const session = await authInstance.api.getSession({ headers: c.req.raw.headers });
-    if (!session) return c.json({ success: false, error: "请先登录" }, 401);
-
-    const teamId = c.req.param("id");
-    const db = createDb(c.env.DB);
-
-    const team = await db.query.teams.findFirst({ where: eq(schema.teams.id, teamId) });
-    if (!team) return c.json({ success: false, error: "队伍不存在" }, 404);
-    if (team.status !== "formed") return c.json({ success: false, error: "只有已组建的队伍需要申请退出" }, 400);
-
-    const members = await db.query.teamMembers.findMany({
-      where: and(eq(schema.teamMembers.teamId, teamId), eq(schema.teamMembers.userId, session.user.id), eq(schema.teamMembers.status, "approved")),
-      limit: 1,
-    });
-    const membership = members[0];
-    if (!membership) return c.json({ success: false, error: "你不是该队伍成员" }, 400);
-    if (membership.userId === team.leaderId) return c.json({ success: false, error: "队长不能退出队伍" }, 400);
-
-    const leaveRequests = await db.query.teamMembers.findMany({
-      where: and(eq(schema.teamMembers.teamId, teamId), eq(schema.teamMembers.userId, session.user.id), eq(schema.teamMembers.status, "leave_pending")),
-      limit: 1,
-    });
-    if (leaveRequests.length > 0) return c.json({ success: false, error: "您已提交退出申请，请等待队长审批" }, 400);
-
-    await db.update(schema.teamMembers)
-      .set({ status: "leave_pending" })
-      .where(eq(schema.teamMembers.id, membership.id));
-
-    return c.json({ success: true, message: "退出申请已提交，等待队长审批" });
-  } catch (error) {
-    console.error("Request leave error:", error);
-    return c.json({ success: false, error: "提交退出申请失败" }, 500);
-  }
-});
-
-/**
- * POST /teams/:id/members/:userId/approve-leave
- * 批准退出申请（仅队长）
- */
-teams.post("/:id/members/:userId/approve-leave", async (c) => {
-  try {
-    const authInstance = createAuth(c.env);
-    const session = await authInstance.api.getSession({ headers: c.req.raw.headers });
-    if (!session) return c.json({ success: false, error: "请先登录" }, 401);
-
-    const teamId = c.req.param("id");
-    const targetUserId = c.req.param("userId");
-    const db = createDb(c.env.DB);
-
-    const team = await db.query.teams.findFirst({ where: eq(schema.teams.id, teamId) });
-    if (!team) return c.json({ success: false, error: "队伍不存在" }, 404);
-    if (team.leaderId !== session.user.id) return c.json({ success: false, error: "只有队长可以批准退出申请" }, 403);
-
-    const members = await db.query.teamMembers.findMany({
-      where: and(eq(schema.teamMembers.teamId, teamId), eq(schema.teamMembers.userId, targetUserId), eq(schema.teamMembers.status, "leave_pending")),
-      limit: 1,
-    });
-    const membership = members[0];
-    if (!membership) return c.json({ success: false, error: "未找到该成员的退出申请" }, 404);
-
-    await db.delete(schema.teamMembers).where(eq(schema.teamMembers.id, membership.id));
-
-    const [{ remainingCount }] = await db
-      .select({ remainingCount: sql<number>`count(*)` })
-      .from(schema.teamMembers)
-      .where(and(eq(schema.teamMembers.teamId, teamId), eq(schema.teamMembers.status, "approved")));
-
-    await db.update(schema.teams)
-      .set({ status: remainingCount < team.maxMembers ? "recruiting" : team.status, updatedAt: new Date() })
-      .where(eq(schema.teams.id, teamId));
-
-    return c.json({ success: true, message: "已批准退出申请" });
-  } catch (error) {
-    console.error("Approve leave error:", error);
-    return c.json({ success: false, error: "批准退出申请失败" }, 500);
-  }
-});
-
-/**
- * POST /teams/:id/members/:userId/reject-leave
- * 拒绝退出申请（仅队长）
- */
-teams.post("/:id/members/:userId/reject-leave", async (c) => {
-  try {
-    const authInstance = createAuth(c.env);
-    const session = await authInstance.api.getSession({ headers: c.req.raw.headers });
-    if (!session) return c.json({ success: false, error: "请先登录" }, 401);
-
-    const teamId = c.req.param("id");
-    const targetUserId = c.req.param("userId");
-    const db = createDb(c.env.DB);
-
-    const team = await db.query.teams.findFirst({ where: eq(schema.teams.id, teamId) });
-    if (!team) return c.json({ success: false, error: "队伍不存在" }, 404);
-    if (team.leaderId !== session.user.id) return c.json({ success: false, error: "只有队长可以拒绝退出申请" }, 403);
-
-    const members = await db.query.teamMembers.findMany({
-      where: and(eq(schema.teamMembers.teamId, teamId), eq(schema.teamMembers.userId, targetUserId), eq(schema.teamMembers.status, "leave_pending")),
-      limit: 1,
-    });
-    const membership = members[0];
-    if (!membership) return c.json({ success: false, error: "未找到该成员的退出申请" }, 404);
-
-    await db.update(schema.teamMembers)
-      .set({ status: "approved" })
-      .where(eq(schema.teamMembers.id, membership.id));
-
-    return c.json({ success: true, message: "已拒绝退出申请" });
-  } catch (error) {
-    console.error("Reject leave error:", error);
-    return c.json({ success: false, error: "拒绝退出申请失败" }, 500);
-  }
-});
-
-/**
- * GET /teams/:id/my-status
- * 获取当前用户在队伍中的成员状态
- */
-teams.get("/:id/my-status", async (c) => {
-  try {
-    const authInstance = createAuth(c.env);
-    const session = await authInstance.api.getSession({ headers: c.req.raw.headers });
-    if (!session) return c.json({ success: true, status: null });
-
-    const teamId = c.req.param("id");
-    const db = createDb(c.env.DB);
-
-    const members = await db.query.teamMembers.findMany({
-      where: and(eq(schema.teamMembers.teamId, teamId), eq(schema.teamMembers.userId, session.user.id)),
-      limit: 1,
-    });
-    const membership = members[0];
-
-    return c.json({ success: true, status: membership ? membership.status : null });
-  } catch (error) {
-    console.error("Get my status error:", error);
-    return c.json({ error: "获取状态失败" }, 500);
-  }
-});
-
-/**
- * POST /teams/:id/form
- * 组建队伍（仅队长）
- */
-teams.post("/:id/form", async (c) => {
-  try {
-    const authInstance = createAuth(c.env);
-    const session = await authInstance.api.getSession({ headers: c.req.raw.headers });
-    if (!session) return c.json({ success: false, error: "请先登录" }, 401);
-
-    const teamId = c.req.param("id");
-    const db = createDb(c.env.DB);
-
-    const body = await c.req.json<{ isUnderfilled?: boolean }>().catch(() => ({} as { isUnderfilled?: boolean }));
-    const isUnderfilled = body.isUnderfilled === true;
-
-    const team = await db.query.teams.findFirst({ where: eq(schema.teams.id, teamId) });
-    if (!team) return c.json({ success: false, error: "队伍不存在" }, 404);
-    if (team.leaderId !== session.user.id) return c.json({ success: false, error: "只有队长可以组建队伍" }, 403);
-    if (team.status !== "recruiting" && team.status !== "full")
-      return c.json({ success: false, error: "当前队伍状态无法组建" }, 400);
-
-    const [{ approvedCount }] = await db
-      .select({ approvedCount: sql<number>`count(*)` })
-      .from(schema.teamMembers)
-      .where(and(eq(schema.teamMembers.teamId, teamId), eq(schema.teamMembers.status, "approved")));
-
-    if (approvedCount < 1) return c.json({ success: false, error: "队伍至少需要1人才能组建" }, 400);
-
-    await db.update(schema.teams)
-      .set({ status: "formed", updatedAt: new Date() })
-      .where(eq(schema.teams.id, teamId));
-
-    return c.json({ success: true, message: "队伍已组建", isUnderfilled });
-  } catch (error) {
-    console.error("Form team error:", error);
-    return c.json({ success: false, error: "组建队伍失败" }, 500);
-  }
-});
-
-/**
- * POST /teams/:id/cancel
- * 取消队伍（仅队长，仅 recruiting/full 状态）
- */
-teams.post("/:id/cancel", async (c) => {
-  try {
-    const authInstance = createAuth(c.env);
-    const session = await authInstance.api.getSession({ headers: c.req.raw.headers });
-    if (!session) return c.json({ success: false, error: "请先登录" }, 401);
-
-    const teamId = c.req.param("id");
-    const db = createDb(c.env.DB);
-
-    const team = await db.query.teams.findFirst({ where: eq(schema.teams.id, teamId) });
-    if (!team) return c.json({ success: false, error: "队伍不存在" }, 404);
-    if (team.leaderId !== session.user.id)
-      return c.json({ success: false, error: "只有队长可以取消队伍" }, 403);
-    if (team.status !== "recruiting" && team.status !== "full")
-      return c.json({ success: false, error: "当前队伍状态无法取消" }, 400);
-
-    await db.update(schema.teams)
-      .set({ status: "cancelled", updatedAt: new Date() })
-      .where(eq(schema.teams.id, teamId));
-
-    return c.json({ success: true, message: "队伍已取消" });
-  } catch (error) {
-    console.error("Cancel team error:", error);
-    return c.json({ success: false, error: "取消队伍失败" }, 500);
-  }
-});
-
-/**
- * DELETE /teams/:id
- * 删除队伍（仅队长，仅 recruiting/cancelled 状态）
- */
-teams.delete("/:id", async (c) => {
-  try {
-    const authInstance = createAuth(c.env);
-    const session = await authInstance.api.getSession({ headers: c.req.raw.headers });
-    if (!session) return c.json({ success: false, error: "请先登录" }, 401);
-
-    const teamId = c.req.param("id");
-    const userId = session.user.id;
-    const db = createDb(c.env.DB);
-
-    const team = await db.query.teams.findFirst({ where: eq(schema.teams.id, teamId) });
-    if (!team) return c.json({ success: false, error: "队伍不存在" }, 404);
-
-    if (team.leaderId !== userId)
-      return c.json({ success: false, error: "只有队长可以删除队伍" }, 403);
-
-    if (team.status !== "recruiting" && team.status !== "cancelled")
-      return c.json({ success: false, error: "只有招募中或已取消的队伍可以删除" }, 400);
-
-    await db.delete(schema.teamMembers).where(eq(schema.teamMembers.teamId, teamId));
-    await db.delete(schema.teams).where(eq(schema.teams.id, teamId));
-
-    return c.json({ success: true, message: "队伍已删除" });
-  } catch (error) {
-    console.error("Delete team error:", error);
-    return c.json({ success: false, error: "删除队伍失败" }, 500);
-  }
-});
-
-async function notifyLeaderOfApplication(
-  db: ReturnType<typeof createDb>,
-  team: typeof schema.teams.$inferSelect,
-  applicantUserId: string,
-  env: Env,
-) {
-  const [leaderRow, applicantRow, locationRow] = await Promise.all([
-    db.select({ email: schema.users.email, name: schema.users.name, nickname: schema.users.nickname })
-      .from(schema.users).where(eq(schema.users.id, team.leaderId)).then((r) => r[0]),
-    db.select({ name: schema.users.name, nickname: schema.users.nickname })
-      .from(schema.users).where(eq(schema.users.id, applicantUserId)).then((r) => r[0]),
-    db.select({ name: schema.locations.name })
-      .from(schema.locations).where(eq(schema.locations.id, team.locationId)).then((r) => r[0]),
-  ]);
-
-  if (!leaderRow || !applicantRow || !locationRow) return;
-
-  const frontendUrl = env.FRONTEND_URL || "https://gomate.live";
-  await sendTeamJoinApplicationEmail(
-    {
-      leaderEmail: leaderRow.email,
-      leaderName: leaderRow.nickname || leaderRow.name,
-      applicantName: applicantRow.nickname || applicantRow.name,
-      teamTitle: team.title,
-      locationName: locationRow.name,
-      teamUrl: `${frontendUrl}/teams/${team.id}`,
-    },
-    env,
-  );
-}
-
+// This file re-exports from the teams/ directory for backward compatibility
+// The teams routes have been split into multiple files:
+// - teams/queries.ts    - GET endpoints
+// - teams/mutations.ts - POST /, PUT /:id, DELETE /:id, POST /:id/form, POST /:id/cancel
+// - teams/membership.ts - POST /:id/join, POST /:id/members/:userId/*, POST /:id/leave-request
+// - teams/status.ts     - POST /:id/leave, POST /:id/cancel-application, POST /:id/members/:userId/approve-leave, POST /:id/members/:userId/reject-leave
+// - teams/utils.ts      - Shared utility functions
+// - teams/index.ts      - Main router that combines all routes
+
+import teams from "./teams/index";
+
+export { teams };
 export { teams as teamsRoute };
+export { default } from "./teams/index";
